requirements.txt
openai
botbuilder-integration-aiohttp>=4.14.0
pillow
azure-ai-projects
azure-identity
python-dotenv
jsonref
opencensus-ext-azure
